pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "cryptography>=46.0.1",
    "fastapi>=0.116.2",
    "flet[all]==0.28.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "websocket-client[optional]>=1.9.0",
    "websockets>=15.0.1",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
